+++ CHANGELOG.md
@@ -30,6 +30,15 @@ You can check your current version with the following command:
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
 
+## 2.1.0a4
+
+**Sep 11, 2024**
+- `Order.get_assets` now allows to get assets from orders in `BEING_FULFILLED` state.
+- Switch `Order` class to use `session` descriptor.
+- Set `OrderParams` as input type for `CatalogBase::place` and `Catalog::estimate` from `Optional[dict]`.
+- Added `OrderParamsV2` as output type for `_translate_construct_parameters` in order module.
+- Changed `Order::status` type from `str` to `Literal`.
+
 ## 2.1.0a3
 
 **Sep 10, 2024**
@@ -51,7 +60,7 @@ For more information, see [UP42 Python package description](https://pypi.org/pro
 
 **Aug 15, 2024**
 
-- Switch `ProductGlossart::IntegrationValue` from `Enum` to `Literal`.
+- Switch `ProductGlossary::IntegrationValue` from `Enum` to `Literal`.
 - Fixed `FEASIBILITY_STUDY_MAY_BE_REQUIRED` value in the `glossary.IntegrationValue` Literal.
 
 
@@ -72,7 +81,7 @@ For more information, see [UP42 Python package description](https://pypi.org/pro
 
 **Aug 13, 2024**
 
-- Switch `ProductGlossart::IntegrationValue` from `Enum` to `Literal`.
+- Switch `ProductGlossary::IntegrationValue` from `Enum` to `Literal`.
 
 ## 2.0.0